# docs/openapi/solvent.yaml
openapi: "3.1.0"
info:
  title: Solvent Authorization Kernel API
  version: "1.0.0"
  description: |
    Transactional belief ledger for autonomous agents. The API is a frozen
    canonical surface: no new endpoints, no shape changes, no behavioral
    changes. Phase 4A remains the original semantic source except where an
    explicitly documented security decision superseded it.

    Authorization model:
    - Beliefs carry evidence, evolve, may be promoted or retracted.
    - Authority is gated by CockroachDB invariants, not application logic.
    - The database — not the API — determines whether an action is allowed.

    Forbidden endpoints (Phase 4A §8): The following patterns are
    structurally prohibited and will never be added:
    /execute-without-authority, /set-authorized, /approve-with-agent,
    /update-authority-state, /force-execution, /override-authorization,
    /set-actor, /impersonate, /approve-and-execute, /execute-as-approved,
    /trust-tool-output, /cache-authorization.

  contact:
    name: Pithom Labs
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /v1/authorizations/execute:
    post:
      operationId: executeAction
      summary: Execute an authorized action (claim intent, invoke provider, record outcome)
      description: |
        The execution boundary. Claims a live intent, invokes the configured
        executor with the approved snapshot parameters, and records the outcome.

        Security invariants:
        - Authenticated principal is server-derived, never caller-supplied.
        - Snapshot consequence_parameters are read from the database, not from the caller.
        - The executor is resolved from the internal registry, not caller-supplied.
        - intent_id is the authoritative execution identity.
      tags: [Authorizations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteActionRequest"
      responses:
        "200":
          description: Execution attempted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecuteActionResult"
        "403":
          description: |
            Forbidden. Possible codes:
            - `actor_id_mismatch`: actor_id does not match authenticated principal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: Execution service not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"

  # ──────────────────────────────────────────────
  # Discharge
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API key passed as Bearer token. The token maps to a single
        principal_id. The authenticated principal is the effective identity
        for all server-derived fields.
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
          examples:
            missing:
              summary: Missing Authorization header
              value:
                code: missing_authorization
                message: "missing_authorization"
                retryable: false
            invalid:
              summary: Invalid token
              value:
                code: invalid_token
                message: "invalid_token"
                retryable: false
...
  schemas:
...
    APIError:
      type: object
      required: [code, message, retryable]
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: validation_error
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Optional structured error details
        retryable:
          type: boolean
          description: Whether the request is safe to retry
        request_id:
          type: string
          description: Optional request identifier for tracing
        sqlstate:
          type: string
          description: PostgreSQL SQLSTATE code if the error originated from a DB constraint
        constraint_name:
          type: string
          description: Database constraint name if applicable

    # ────────────────────────────────────────────
    # Beliefs
    # ────────────────────────────────────────────
...
    ExecuteActionRequest:
      type: object
      required: [scenario_id, belief_id, action, target_id, intent_id]
      description: |
        Execute an authorized action. The authenticated principal is derived
        from the API key, not from the request body. consequence_parameters
        are read from the approved snapshot, not from the caller.
      properties:
        scenario_id:
          type: string
          format: uuid
          description: Scenario containing the belief
        belief_id:
          type: string
          format: uuid
          description: Promoted belief authorizing the action
        action:
          type: string
          description: Action to execute (must match the authorized action)
        target_id:
          type: string
          format: uuid
          description: Authority target with approved snapshot
        intent_id:
          type: string
          format: uuid
          description: Live intent to claim and execute
        consequence_type:
          type: string
          description: Consequence type (defaults to "execution")
    ExecuteActionResult:
      type: object
      properties:
        belief_id:
          type: string
          format: uuid
        intent_id:
          type: string
          format: uuid
        intent_state:
          type: string
          description: 'executing if provider accepted but not completed, executed if completed'

        action:
          type: string
        allowed:
          type: boolean
          description: Whether kernel.Authorize granted authority
        success:
          type: boolean
          description: Whether the executor completed successfully
        output:
          type: string
          description: Provider output (e.g., GitHub run ID)
        error:
          type: string
          description: Error message if execution failed
        reason:
          type: string
          description: Denial reason if authority was not granted
        executed_at:
          type: string
          format: date-time

    # ────────────────────────────────────────────
    # Discharge
    # ────────────────────────────────────────────
...